Fix for TestCaseImport.cy.ts
